% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggimprensa-package.R
\docType{package}
\name{ggimprensa-package}
\alias{ggimprensa}
\alias{ggimprensa-package}
\title{ggimprensa: Temas dos principais jornais brasileiros para o ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Faça gráficos profissionais inspirtados nos jornais brasileiros.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/ggimprensa}
  \item Report bugs at \url{https://github.com/brunomioto/ggimprensa/issues}
}

}
\author{
\strong{Maintainer}: Bruno Mioto \email{brunomioto97@gmail.com} (\href{https://orcid.org/0000-0002-3524-537X}{ORCID})

}
\keyword{internal}
