% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_poder360.R
\name{linha_poder360}
\alias{linha_poder360}
\title{Adiciona linha superior}
\usage{
linha_poder360(plot)
}
\arguments{
\item{plot}{O grafico criado anteriormente com a funcao tema_poder360()}
}
\description{
Adiciona linha superior
}
