% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_folha.R
\name{scale_fill_folha}
\alias{scale_fill_folha}
\title{Escala de preenchimento para a paleta de cores Folha}
\usage{
scale_fill_folha(..., na.value = "grey50")
}
\arguments{
\item{...}{Argumentos adicionais passados para \code{discrete_scale}.}

\item{na.value}{A cor a ser usada para valores ausentes. O valor padrão é \code{"grey50"}.}
}
\value{
Uma função de escala do ggplot2 que aplica a paleta de cores Folha.
}
\description{
Uma escala personalizada do ggplot2 para a estética de `fill` usando a paleta de cores Folha.
A paleta é composta por um conjunto de nove cores distintas,
projetadas para serem usadas em variáveis discretas de preenchimento em gráficos do ggplot2.
}
