% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_folha.R
\name{tema_folha}
\alias{tema_folha}
\title{Tema inspirado nos graficos do jornal Folha de Sao Paulo}
\usage{
tema_folha(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Tema inspirado nos graficos do jornal Folha de Sao Paulo
}
\note{
Fontes originais

 Titulo e subtitulo: FolhaGrafico

 Grafico: FolhaGrafico
}
\concept{themes folha}
