% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_g1.R
\name{tema_g1}
\alias{tema_g1}
\title{Tema inspirado nos graficos do G1 - globo.com}
\usage{
tema_g1(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Tema inspirado nos graficos do G1 - globo.com
}
\note{
Fontes originais

 Titulo e subtitulo:

 Grafico:
}
\concept{themes g1}
