% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_nexo.R
\name{tema_nexo}
\alias{tema_nexo}
\title{Tema inspirado nos graficos do Nexo Jornal}
\usage{
tema_nexo(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Tema inspirado nos graficos do Nexo Jornal
}
\note{
Fontes originais

 Titulo e subtitulo: Sentinel

 Grafico: Gotham Rounded
}
\concept{themes nexo}
