% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_poder360.R
\name{tema_poder360}
\alias{tema_poder360}
\title{Tema inspirado nos graficos do jornal Poder 360}
\usage{
tema_poder360(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Tema inspirado nos graficos do jornal Poder 360
}
\concept{themes poder360}
