% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_nexo.R
\name{tema_nexo}
\alias{tema_nexo}
\title{Tema inspirado nos gráficos do Nexo Jornal}
\usage{
tema_nexo(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Tema inspirado nos gráficos do Nexo Jornal
}
\note{
Fontes originais:
 Título e subtítulo: Sentinel
 Gráfico: Gotham Rounded
}
\concept{themes nexo}
